fix(migrations): move operations-log and provenance rules into base schema
- Push operations-log fields and provenance constraints into the initial schema definitions.
- Remove the need for later patch migrations to reach the shape runtime code already expects.
- Keep fresh installs closer to upgraded installs.

+++ crate/lib/sinex-core/src/db/repositories/knowledge_graph.rs
@@ -581,7 +581,7 @@ impl<'a> KnowledgeGraphRepository<'a> {
             r#"
             INSERT INTO core.entity_relations (
                 id, from_entity_id, to_entity_id, relation_type, 
-                strength as f32, metadata, valid_from, valid_until, created_from_event_id
+                strength, metadata, valid_from, valid_until, created_from_event_id
             ) VALUES (
                 $1, $2, $3, $4, $5, $6, $7, $8, $9
             )
